+++ lib/src/widgets/base_phone_form_field.dart
@@ -3,6 +3,7 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:phone_form_field/src/models/simple_phone_number.dart';
 
+import '../../phone_form_field.dart';
 import '../models/country.dart';
 import 'country_picker/country_selector_navigator.dart';
 import 'flag_dial_code_chip.dart';