+++ CHANGELOG.md
@@ -11,6 +11,11 @@
 - Country code is now visible when unfocussed when there is no label to better
   accomodate the UI
 
+# Validation
+- Add PhoneValidator class to easily customize validation and defaults localization error messagees
+- Add PhoneFormField `validator` property
+- **[BREAKING CHANGE]** Remove `PhoneFormField` properties `errorText` & `phoneNumberType`. Define `validator` property instead with `PhoneValidator.invalid*`
+
 # Misc
 - refactored internals